Split retained stdlib glue from manifest
