+++ crates/sifr_lowering/Cargo.toml
@@ -12,7 +12,8 @@ sifr_ir = { workspace = true }
 sifr_ipc = { workspace = true }
 sifr_python_ast = { workspace = true }
 sifr_python_parser = { workspace = true }
-sifr_stdlib_manifest = { workspace = true }
+sifr_retained_intrinsics = { workspace = true }
+sifr_stdlib_imports = { workspace = true }
 sifr_type_system = { workspace = true }
 ruff_text_size = { workspace = true }
 rust_decimal = "1.41.0"
