+++ crates/sifr_retained_intrinsics/src/crypto_regex_uuid.rs
