+++ crates/sifr_retained_intrinsics/src/http.rs
