+++ crates/sifr_retained_intrinsics/src/i18n_core.rs
