+++ crates/sifr_retained_intrinsics/src/io_json.rs
