+++ crates/sifr_retained_intrinsics/src/platform_misc.rs
