+++ crates/sifr_retained_intrinsics/src/signal.rs
