+++ crates/sifr_retained_intrinsics/src/text_encoding.rs
