+++ crates/sifr_retained_intrinsics/src/unicode_core.rs
