+++ internal_docs/architecture.md
@@ -272,7 +272,9 @@ sifr/
     sifr_diagnostics/       (canonical diagnostic codes, source-map spans, model, render schema, and sink)
     sifr_ir/                (High-level IR data rules, public lowered views, CFG/flow graph data)
     sifr_lowering/          (AST-to-IR lowering, name/type/ownership/async analysis, lowering diagnostics)
-    sifr_stdlib_manifest/   (compiler-host stdlib source inventory, private declaration inventory, feature planning, sysroot validation, and import suggestions; retained-signature responsibilities are being narrowed into a dedicated boundary)
+    sifr_stdlib_manifest/   (compiler-host stdlib source inventory, private declaration inventory, feature planning, and sysroot validation)
+    sifr_stdlib_imports/    (legacy and bare stdlib import suggestion policy over manifest inventory)
+    sifr_retained_intrinsics/ (temporary compiler-retained stdlib intrinsic signatures)
     sifr_ipc/               (shared IPC protocol/frame/schema/request-tracking crate)
     sifr_type_system/       (type definitions, inference, checking, subtyping)
     sifr_codegen/           (Rust source code generation from HIR via structured Rust IR)
