+++ plans/reviews/active/stdlib-native-boundary-m0c-retained-glue-review-pass1.md
@@ -1,0 +1,1 @@
+**Verdict: PR is ready to merge.** The one stale line in `network_http_baseline_traceability.md:16` is a low-severity doc consistency nit (the three sibling reports were updated correctly), and the two unused predicates in `sifr_retained_intrinsics/src/lib.rs` are inherited from the pre-move manifest and out of M0c's mechanical-move scope. No dependency-direction regression, no lost test, no missing enforcement, no behavioral change in import diagnostics or intrinsic bootstrap resolution.
