+++ plans/reviews/active/stdlib-native-boundary-m0c-retained-glue-review-pass2.md
@@ -1,0 +1,21 @@
+## Pass-2 findings
+
+No actionable findings.
+
+## Verification against pass-2 focus areas
+
+- **`sifr_stdlib_manifest` is metadata/planning only.** `crates/sifr_stdlib_manifest/src/lib.rs:11-22` re-exports only feature planning (`StdlibFeature`, `STDLIB_FEATURE_SPECS`, `SysrootDependencyPlan`, `try_generated_cargo_dependencies`, …) and source inventory (`STDLIB_SOURCES`, `PRIVATE_STDLIB_MODULES`, `load_stdlib_sources_from_sysroot`, …). The 18 per-family intrinsic files are deleted from the manifest crate. The manifest Cargo.toml drops `sifr_type_system` (`crates/sifr_stdlib_manifest/Cargo.toml:10`). Grep confirms zero `sifr_type_system` refs remain under `crates/sifr_stdlib_manifest/`.
+- **Import suggestion policy owned by `sifr_stdlib_imports`.** `crates/sifr_stdlib_imports/src/lib.rs` owns `BareStdlibMatch`, `LegacyStdlibModule`, `is_bare_stdlib_tail`, `unsupported_legacy_stdlib_module`, and the CPython-shaped reserved-suggestion table. It queries manifest only via `STDLIB_SOURCES` for existence lookups (lines 213–223). Callers (`sifr_lowering`, `sifr_driver::project::{discovery, package_discovery}`) all switched to the new crate.
+- **Retained signature builders owned by `sifr_retained_intrinsics`.** `crates/sifr_retained_intrinsics/src/lib.rs` hosts `IntrinsicModule`, `get_intrinsic_module`, and all 19 per-module builder submodules. Depends only on `sifr_type_system`. Consumers: `sifr_lowering::lower::mod_impl` (line 278) and `sifr_driver::stdlib::bootstrap` (lines 200/548/626). No lowering→codegen edge introduced.
+- **Dependency direction guard coherence.** `scripts/check_source_crate_dependency_direction.py`:
+  - Adds both crates to `ALL_SIFR_CRATES` and to the `IR_FORBIDDEN_DEPENDENCIES`, `STDLIB_FORBIDDEN_DEPENDENCIES`, and `GENERATED_STDLIB_FORBIDDEN_DEPENDENCIES` sets.
+  - Adds `CrateRule` entries with strict allowlists (`sifr_retained_intrinsics` → only `sifr_type_system`; `sifr_stdlib_imports` → only `sifr_stdlib_manifest`).
+  - Adds four self-tests exercising both "unexpected normal dependency" and "source reference" failure paths (`scripts/check_source_crate_dependency_direction.py:405-441`).
+  - `seed_valid_repo` seeds both crates so the self-test's clean baseline still passes.
+- **Validation profiles, coverage metadata, generated-code-quality fingerprints.** `verification/profiles/{create-pr,merge,nightly,release}.json` all add `sifr_retained_intrinsics` and `sifr_stdlib_imports` merge-blocking suites. `verification/areas/coverage_matrix/data/cargo_metadata_classification.json` adds `first_party_compiler` entries for both. `verification/areas/generated_code_quality/generated_code_quality.py` adds both to `PRODUCER_FINGERPRINT_CRATES`.
+- **Docs coherence.** `internal_docs/architecture.md:274-277` and `internal_docs/sifr_sysroot_and_stdlib_architecture.md:59,621-638` describe the new three-crate split. `internal_docs/hir_maintainability_guardrails.md:3-5,41` updated to reflect that manifest, import policy, and retained signatures live in separate crates. Remaining `sifr_stdlib_manifest::` references (in codegen, text_i18n/network_http architecture docs, dependency-snapshot data) all point to legitimate manifest concerns (feature planning, generated Cargo, snapshot tests) — not stale ownership claims.
+- **Pass-1 doc nit.** `verification/areas/stdlib_parity/reports/network_http_baseline_traceability.md:16` now reads `crates/sifr_stdlib_imports/src/lib.rs`. The three sibling reports (`network_http_tls_traceability.md`, `network_http_url_header_cookie_traceability.md`, `concurrency_runtime_legacy_surface_traceability.md`) were already updated in the primary diff.
+
+## Verdict
+
+**PR is ready to merge.** No blocking or actionable issues. The `sifr_retained_intrinsics::is_intrinsic_module`/`is_stdlib_module` predicates that pass-1 flagged remain unused externally, but they are pre-existing public API surface inherited verbatim from the pre-move manifest, and are out of M0c's mechanical-move scope — appropriate to leave for a later cleanup milestone when the retained-glue crate shrinks toward M13 closure.
