+++ scripts/check_source_crate_dependency_direction.py
@@ -33,9 +33,11 @@ ALL_SIFR_CRATES = {
     "sifr_lowering",
     "sifr_lsp",
     "sifr_package",
+    "sifr_retained_intrinsics",
     "sifr_runtime",
     "sifr_source",
     "sifr_stdlib",
+    "sifr_stdlib_imports",
     "sifr_stdlib_manifest",
     "sifr_sysroot",
     "sifr_syntax",
@@ -53,7 +55,9 @@ IR_FORBIDDEN_DEPENDENCIES = {
     "sifr_lowering",
     "sifr_lsp",
     "sifr_package",
+    "sifr_retained_intrinsics",
     "sifr_stdlib",
+    "sifr_stdlib_imports",
     "sifr_stdlib_manifest",
     "sifr_syntax",
 }
@@ -69,6 +73,8 @@ STDLIB_FORBIDDEN_DEPENDENCIES = {
     "sifr_lsp",
     "sifr_package",
     "sifr_ipc",
+    "sifr_retained_intrinsics",
+    "sifr_stdlib_imports",
     "sifr_stdlib_manifest",
 }
 
@@ -84,6 +90,8 @@ GENERATED_STDLIB_FORBIDDEN_DEPENDENCIES = {
     "sifr_lowering",
     "sifr_lsp",
     "sifr_package",
+    "sifr_retained_intrinsics",
+    "sifr_stdlib_imports",
     "sifr_stdlib_manifest",
     "sifr_syntax",
     "sifr_sysroot",
@@ -131,6 +139,20 @@ RULES = (
         crate="sifr_ipc",
         allowed_normal_dependencies=frozenset({"postcard", "serde"}),
         forbidden_source_references=frozenset(ALL_SIFR_CRATES - {"sifr_ipc"}),
+    ),
+    CrateRule(
+        crate="sifr_retained_intrinsics",
+        allowed_normal_dependencies=frozenset({"sifr_type_system"}),
+        forbidden_source_references=frozenset(
+            ALL_SIFR_CRATES - {"sifr_retained_intrinsics", "sifr_type_system"}
+        ),
+    ),
+    CrateRule(
+        crate="sifr_stdlib_imports",
+        allowed_normal_dependencies=frozenset({"sifr_stdlib_manifest"}),
+        forbidden_source_references=frozenset(
+            ALL_SIFR_CRATES - {"sifr_stdlib_imports", "sifr_stdlib_manifest"}
+        ),
     ),
     CrateRule(
         crate="sifr_stdlib_manifest",
@@ -298,8 +320,9 @@ def seed_valid_repo(root: Path) -> None:
         "sifr_source": ["ruff_text_size"],
         "sifr_ir": ["sifr_diagnostics", "sifr_type_system"],
         "sifr_ipc": ["postcard", "serde"],
+        "sifr_retained_intrinsics": ["sifr_type_system"],
         "sifr_stdlib": ["sifr_runtime"],
-        "sifr_stdlib_manifest": ["sifr_type_system"],
+        "sifr_stdlib_imports": ["sifr_stdlib_manifest"],
         "sifr_codegen": ["sifr_ir", "sifr_stdlib_manifest"],
         "sifr_lint": ["sifr_frontend", "sifr_ir"],
         "sifr_analysis": ["sifr_frontend", "sifr_lint"],
@@ -380,6 +403,42 @@ def run_self_test() -> int:
             root / "crates" / "sifr_ipc" / "src" / "lib.rs"
         ).write_text("pub fn leak() { sifr_stdlib_manifest::STDLIB_FEATURE_SPECS; }\n", encoding="utf-8"),
         "sifr_ipc: crates/sifr_ipc/src/lib.rs references sifr_stdlib_manifest",
+        failures,
+    )
+    assert_self_test_case(
+        "sifr_retained_intrinsics compiler dependency",
+        lambda root: write_manifest(
+            root / "crates" / "sifr_retained_intrinsics",
+            "sifr_retained_intrinsics",
+            ["sifr_stdlib_manifest"],
+        ),
+        "sifr_retained_intrinsics: unexpected normal dependency",
+        failures,
+    )
+    assert_self_test_case(
+        "sifr_retained_intrinsics source reference",
+        lambda root: (
+            root / "crates" / "sifr_retained_intrinsics" / "src" / "lib.rs"
+        ).write_text("pub fn leak() { sifr_stdlib_manifest::STDLIB_FEATURE_SPECS; }\n", encoding="utf-8"),
+        "sifr_retained_intrinsics: crates/sifr_retained_intrinsics/src/lib.rs references sifr_stdlib_manifest",
+        failures,
+    )
+    assert_self_test_case(
+        "sifr_stdlib_imports compiler dependency",
+        lambda root: write_manifest(
+            root / "crates" / "sifr_stdlib_imports",
+            "sifr_stdlib_imports",
+            ["sifr_driver"],
+        ),
+        "sifr_stdlib_imports: unexpected normal dependency",
+        failures,
+    )
+    assert_self_test_case(
+        "sifr_stdlib_imports source reference",
+        lambda root: (
+            root / "crates" / "sifr_stdlib_imports" / "src" / "lib.rs"
+        ).write_text("pub fn leak() { sifr_driver::compile(); }\n", encoding="utf-8"),
+        "sifr_stdlib_imports: crates/sifr_stdlib_imports/src/lib.rs references sifr_driver",
         failures,
     )
     assert_self_test_case(
