+++ verification/areas/stdlib_parity/reports/concurrency_runtime_legacy_surface_traceability.md
@@ -18,7 +18,7 @@ Implementation evidence:
 
 - `lib/sifr/asyncio.sifr`, `lib/sifr/concurrent.sifr`, `lib/sifr/subprocess.sifr`, and `lib/sifr/threading.sifr` were removed.
 - `crates/sifr_stdlib_manifest/src/sources.rs` no longer embeds those legacy modules.
-- `sifr_stdlib_manifest::unsupported_legacy_stdlib_module` records native replacement namespaces.
+- `sifr_stdlib_imports::unsupported_legacy_stdlib_module` records native replacement namespaces.
 - `SIFR-IMPORT-0009` distinguishes removed legacy Sifr modules from unknown modules.
 - Legacy `sifr.asyncio` compatibility lowering, `asyncio.run` entrypoint inference, and `LowerCtx.asyncio_compat_imports` were removed so native task lowering no longer depends on CPython-shaped imports.
 - Historical fail fixture names that referenced member-level legacy behavior now assert the module-level removal diagnostic and provide imported-member payload coverage.
