+++ verification/areas/stdlib_parity/reports/network_http_tls_traceability.md
@@ -6,7 +6,7 @@ TLS capability implements the Sifr-native TLS substrate over the async-network c
 
 | Requirement | network/HTTP baseline capability decision | Implementation evidence |
 | --- | --- | --- |
-| `sifr.tls` public module and config types | `production-public`, stable public Sifr API; Rustls types hidden. | `lib/sifr/tls.sifr`; private `_sifr.tls` intrinsics in `crates/sifr_stdlib_manifest/src/tls.rs`; runtime module gate in `crates/sifr_runtime/src/lib.rs`; codegen lowerers in `crates/sifr_codegen/src/intrinsics/registry/tls.rs` and `crates/sifr_codegen/src/preamble/tls_runtime.rs`. |
+| `sifr.tls` public module and config types | `production-public`, stable public Sifr API; Rustls types hidden. | `lib/sifr/tls.sifr`; private `_sifr.tls` intrinsics in `crates/sifr_retained_intrinsics/src/tls.rs`; runtime module gate in `crates/sifr_runtime/src/lib.rs`; codegen lowerers in `crates/sifr_codegen/src/intrinsics/registry/tls.rs` and `crates/sifr_codegen/src/preamble/tls_runtime.rs`. |
 | Safe client verification defaults | Platform verification is the default production client strategy; no fallback root store. | `client_config_platform(...)` uses `rustls-platform-verifier`; dependency snapshot asserts `rustls-platform-verifier = { version = "0.7.0", default-features = false }`; no `webpki-roots` is emitted. |
 | Deterministic test roots | Explicit in-memory roots for tests; `rcgen` is dev/test only. | `client_config_with_roots(...)` and mTLS config helpers parse PEM through `rustls-pemfile`; runtime tests generate deterministic CA/server/client material with dev-only `rcgen`; generated dependency snapshot asserts no `rcgen`, `webpki-roots`, or `x509-parser` in production deps. |
 | TLS client/server streams | Async-native over async-network capability TCP and `tokio-rustls`; TCP handles are consumed into TLS handles. | `crates/sifr_runtime/src/tls.rs`; `crates/sifr_runtime/src/net.rs` exposes internal `consume_stream_for_tls(...)`; public fixture `crates/sifr/tests/e2e/pass/network_http_tls_loopback_split.sifr` performs real loopback client/server handshakes. |
