+++ verification/areas/stdlib_parity/reports/network_http_url_header_cookie_traceability.md
@@ -4,12 +4,12 @@ Status: merged in PR #2497 at `9a3ee4d18a12ab6ddaa9174aebea591a891c4651`.
 
 | Requirement | network/HTTP baseline capability decision | Acceptance evidence |
 | --- | --- | --- |
-| `sifr.url.Url` and `UrlQuery` | `production-public`, stable-public-api over ASCII and already-valid Sifr text. | `lib/sifr/url.sifr`; `_sifr.url` signatures in `crates/sifr_stdlib_manifest/src/url.rs`; generated helpers in `crates/sifr_codegen/src/preamble/url_http_runtime.rs`; fixture `network_http_url_query_percent.sifr`. |
+| `sifr.url.Url` and `UrlQuery` | `production-public`, stable-public-api over ASCII and already-valid Sifr text. | `lib/sifr/url.sifr`; `_sifr.url` signatures in `crates/sifr_retained_intrinsics/src/url.rs`; generated helpers in `crates/sifr_codegen/src/preamble/url_http_runtime.rs`; fixture `network_http_url_query_percent.sifr`. |
 | Percent helpers | Byte/ASCII/UTF-8 safe helpers accepted; named encodings blocked on text/i18n async-network capability. | `percent_encode`, `percent_decode`, `percent_encode_bytes`, and `percent_decode_bytes`; fixture covers UTF-8 text, raw bytes, and invalid percent escapes. |
 | Host and authority validation | ASCII domain labels, IPv4, IPv6, and already-punycode accepted; non-ASCII blocked until text/i18n TLS capability; ports are decimal 0-65535. | Generated helper rejects non-ASCII authority hosts before `url` crate IDNA behavior can leak; fixture covers IPv4 parsing, IPv6 building, already-punycode host, and invalid port rejection. |
 | Path normalization | Parsing applies WHATWG dot-segment removal for special schemes. Percent-encoded slash (`%2F`) is preserved as a segment byte, not a separator. | `normalize_path` helper and fixture coverage for parser dot-segment removal, explicit helper dot-segment removal, and encoded slash boundaries. |
 | Query parsing/building | byte/ASCII/UTF-8 safe behavior only; sensitive query key redaction applies in observability. | `UrlQuery` stores ordered pairs; parse/build fixture covers duplicate keys, plus decoding, and empty/missing lookup behavior. |
-| `sifr.http` header primitives | URL/header/cookie capability owns canonical `HeaderName`, `HeaderValue`, and `HeaderMap`: ASCII token names, lowercase canonical names, obs-fold rejection, OWS trim only, duplicate order preservation, and size caps from inventory. | `lib/sifr/http.sifr`; `_sifr.http` signatures in `crates/sifr_stdlib_manifest/src/http.rs`; generated helpers use `http` crate validation; fixture covers lowercase canonicalization, duplicate order, OWS trim, invalid names, obs-fold rejection, and inventory hard caps. |
+| `sifr.http` header primitives | URL/header/cookie capability owns canonical `HeaderName`, `HeaderValue`, and `HeaderMap`: ASCII token names, lowercase canonical names, obs-fold rejection, OWS trim only, duplicate order preservation, and size caps from inventory. | `lib/sifr/http.sifr`; `_sifr.http` signatures in `crates/sifr_retained_intrinsics/src/http.rs`; generated helpers use `http` crate validation; fixture covers lowercase canonicalization, duplicate order, OWS trim, invalid names, obs-fold rejection, and inventory hard caps. |
 | Cookie header parsing | header-level parse/build only. | Cookie header parse/build uses ordered `list[tuple[str, str]]`; fixture covers parse/build, embedded `=` in values, and invalid cookie syntax. Cookie jars, signing, private cookies, persistence, and percent-decoded user text remain unexposed. |
 
 ## Validation Evidence
